various tweeks, and new genability class.
git-svn-id: svn://192.168.1.10/public/CoffeeMud@24352 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Items/Basic/StdItem.java
@@ -1294,7 +1294,7 @@ public class StdItem implements Item
 				mob.tell(L("@x1 looks too advanced for you.",name(msg.source())));
 				return false;
 			}
-			return canWearComplete(mob,(msg.value()<=0)?0:((long)(1<<msg.value())/2),msg.othersMessage()==null);
+			return canWearComplete(mob,(msg.value()<=0)?0:((1L<<msg.value())/2),msg.othersMessage()==null);
 		case CMMsg.TYP_WIELD:
 			if((!fitsOn(Wearable.WORN_WIELD))||(properWornBitmap==0))
 			{
